test: split runtime plan ledger coverage
